[Syanima/Sooraj/Saran] Added getCurrentPlayer method in GameMaster
Why:

It is game master's responsibility to choose the right player for the play.
Game master allows the current player to play if he has not completed all
his chances .
Game master chooses the next player and gives him one chance to play  if
the current player has completed all his chances.
Then next player will be the current player.

We have written tests for the above functionality but it requires more
tests in the future.