[Sooraj/Syanima] Modified getCurrentPlayer method
What changes have you made?

We have modified a method called 'getCurrentPlayer' in GameMaster.

Why do we need this change?

On game playing forward, we have to know who should play next. If the
last player has not finished his chances, he has to get one more chance
to roll dice and if he lack chance, we need to know the next player.

But the previous version has an issue with the counter. The counter was
not initiated every time while a new gameMaster was created. It was
because of the prototype used for the GameMaster was just reference so
the counter will get initialised only the first time the game master is
created. Further with that, we cannot create multiple game masters in the
server.

What can this change impact?

The method was created as a characteristics in the GameMaster object. So
on every time the game master is created the counter gets initialised
seperately for each object. This can overcome the issue we faced.

We have also writted tests for the above functionality, but has to be
tested for different behaviours later.

*