[Sooraj] Modified killCoin method of Player
What changes have you made?

I have modified the killCoin method of the player entity. An extra line
which increments the chance of the player is added.

Why do we need this change?

When the player is killing a coin the player will get an extra chance to
roll the dice. Then he can play the dice once more after the move.

The whole sequence of the turns lies on the chances of the player. So it
is important to increment and decrement the chances of the player at
exact moment.

What can this change impact?

The added line will increment the chance of the player whenever he kills
a coin. This will impact in the sequence of the game. The player will get
one more chance to roll the dice and play if he kills one coin.