[Supriya/Sooraj] Added get_All_Valid_Moves_Of_Coin
What changes have you made?

We created a function which will give all the possible moves the given
coin can make.

Why do we need this change?

We need to know the all the possible moves, a coin can make, to validate
a move choosen by the player on a particular coin. It is also helpful if
the app needs to show the player all the possible moves he can make as a
hint.

What can this change impact?

It will help in validating the move the player needs to make.

It is necessary to know the path of the player through which the he can
move the coin. But we have not created any method to generate the path
of a particular player. It is mandatory to create a method for the same
and test the above functionality using the method.