[Sooraj/Supriya] - modified getAllValidmoves function.
What changes have you made?

we modified the getAllValidMoves function to check the entering condition.

Why do we need this change?

we need this change to check wheather player's coin is allowed to enter into a board or not.

What can this change impact?

it will help player to check wheather he can enter his/her coin into board or not.