[Saran/Sooraj] Modified and synced DICE button
What changes have you made?

We have linked the DICE button with the roll dice in each player. When
DICE button is pressed, a request is made to the server and the
purticular player rolls the dice.

Why do we need this change?

To play the game, player has to roll the dice and get some value by which
he can move his coins. By clicking the DICE button, only that particular
player should roll the dice.

What can this change impact?

When we click on DICE button, it will roll the dice in server and the
value will be shown as an alert message in our page.