[Sooraj/Syanima] Added path for the player
What changes have you made?

We have made a get method for player which generates the path of the
player according to the maturity.

Why do we need this change?

We need to know the path of the player to move the coin on the board.
And also the is different if the player is matured or un-matured. So
it is necessary to check for the maturity of the player and generate the
path accordingly.

What can this change impact?

We have made another file which has all the functions for generating the
path. The path method generates half path of the player if he is not
matured and also generates full path to the home if the player is
matured.