[Sooraj/Syanima] - modified GeneratePath
What changes have you made?

We have modified a little logical code in generate path file.

Why do we need this change?

As per our old code the path generation was in the reverse order. So
that the path strats from end and goes to front.

Tests for the same is also modified.

What can this change impact?

It is now verified and corrected the path of the coin through which it
has to move.