[Sooraj/Saran] Created roll dice functionality
What changes have you made?

We have created the functionality which enables the player to roll the
dice and know how many steps he can move the coin.

We have also modified the server part as to work with the request.

Why do we need this change?

For the coin movement we have to how many steps we can move. So we need
a dice to give the quantity.

What can this change impact?

When the player will click on dice button, he will get a number between
1 and 6. It is also piled as a list under the dice button.