[Sooraj/Syanima] Created anyMoreMove functionality
What changes have you made?

We have created a method in board entity which returns true or false

Why do we need this change?

We have to know whether the player has any more moves to decide if
player has to move or give out his chance. This is important in
sequencing the player and chance distribution.

What can this change impact?

The written functionality gives out true or false by considering the
dice values, position of coins and the path of the player.