[Sooraj/Syanima] Modified WaitingPage UI
What changes have you made?

* Waiting page is made simple and more informative. Instead of player
count, the waitingPage is now showing the players joined with highlight
to the lastest player joined.

A waiting robin is implemented to have more effect of waiting.

Also server code is changed a little to cope-up the new waiting page.

Why do we need this change?

* As a part of refactoring UI

What can this change impact?

* Everything is working fine.